Use templated functions for hardcoded divtests
This produces the same machine code, but with far less room for programmer error.

Also included, but not used, is a recursive templated div test that automatically works for any prime factor and base, but with a small performance cost. For some reason the compiler produces slightly more code which ends up being slower.

Also:
- update benchmarking notes
- code cleanup