Update benchmark notes - recalculating
Factoring in div test hit counts when sorting div tests results in a ~1.2% speedup. I missed that it provided any performance impact, but left it in the code for further experimenting. As a result, the next optimization "7464 ms : Jump to the next set byte in the sieve" was calculated to provide a 15% speedup from the previous committed version, whereas the actual benefit is closer to 13%. This commit updates the notes to reflect that.

This commit also adds a record, "Evaluate hit counts when sorting div tests". The run time of 8621 ms and the performance impact of 1.2% were calculated as what would have been measured at the time the code was changed.